+++ .eslintignore
@@ -1,2 +1,5 @@
+# temporary files
+coverage
+
 # 3rd-party files
 node_modules
