+++ .gitignore
@@ -1,4 +1,6 @@
 # 3rd-party files
 node_modules
 
+# temporary files
 .eslintcache
+coverage
