+++ __tests__/.eslintrc.json
@@ -1,4 +1,7 @@
 {
+  "env": {
+    "jest": true
+  },
   "rules": {
     "node/no-unpublished-require": 0
   }
